Move HttpStatus to UniversalUtil
See https://github.com/A248/UniversalRegistry
