+++ ArimBansLib/src/main/java/space/arim/bans/api/exception/FetcherException.java
@@ -18,7 +18,7 @@
  */
 package space.arim.bans.api.exception;
 
-import space.arim.bans.api.util.web.HttpStatus;
+import space.arim.registry.util.web.HttpStatus;
 
 public class FetcherException extends InternalAPIException {
